requirements.txt
pygame==2.6.0
python-dotenv==1.0.1
python-telegram-bot==21.4